requirements.txt
asgiref==3.8.1
click==8.2.1
colorama==0.4.6
Django==5.2.3
django-cors-headers==4.9.0
django-debug-toolbar==5.2.0
django-redis==5.4.0
django-rq==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.10.1
python-dotenv==1.1.0
redis==6.2.0
rq==2.3.3
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.9.0
